Move platform DRAM range function to take unshifted 64-bit range

+++ bootloader/dnc-maps.h
@@ -20,7 +20,7 @@
 bool dram_range_read(const uint16_t sci, const int ht, const int range, uint64_t *base, uint64_t *limit, int *dest);
 int dram_range_unused(const uint16_t sci, const int ht);
 void dram_range_print(const uint16_t sci, const int ht, const int range);
-void dram_range(const uint16_t sci, const int ht, const int range, const uint32_t base, const uint32_t limit, const int dest);
+void dram_range(const uint16_t sci, const int ht, const int range, const uint64_t base, const uint64_t limit, const int dest);
 void dram_range_del(const uint16_t sci, const int ht, const int range);
 bool mmio_range_read(const uint16_t sci, const int ht, int range, uint64_t *base, uint64_t *limit, int *dest, int *link, bool *lock);
 void mmio_range_print(const uint16_t sci, const int ht, const int range);
